Refactor: Change name of product list file
